add util of split duration
