+++ iotdb-core/datanode/src/test/java/org/apache/iotdb/db/utils/DateTimeUtilsTest.java
@@ -18,7 +18,7 @@
  */
 package org.apache.iotdb.db.utils;
 
-import org.apache.iotdb.tsfile.utils.duration.TimeDuration;
+import org.apache.iotdb.tsfile.utils.TimeDuration;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -240,26 +240,26 @@ public class DateTimeUtilsTest {
   public void testConstructTimeDuration() {
     TimeDuration timeDuration = DateTimeUtils.constructTimeDuration("1y1d1ns");
     Assert.assertEquals(12, timeDuration.monthDuration);
-    Assert.assertEquals(1000000 * 86400_000L + 1, timeDuration.nsDuration);
+    Assert.assertEquals(86400_000L, timeDuration.currPrecisionDuration);
 
-    timeDuration = DateTimeUtils.constructTimeDuration("1y1mo1d1ns");
+    timeDuration = DateTimeUtils.constructTimeDuration("1y1mo1d1ms1ns");
     Assert.assertEquals(13, timeDuration.monthDuration);
-    Assert.assertEquals(1000000 * 86400_000L + 1, timeDuration.nsDuration);
+    Assert.assertEquals(86400_001L, timeDuration.currPrecisionDuration);
 
     timeDuration = DateTimeUtils.constructTimeDuration("1d1ns");
     Assert.assertEquals(0, timeDuration.monthDuration);
-    Assert.assertEquals(1000000 * 86400_000L + 1, timeDuration.nsDuration);
+    Assert.assertEquals(86400_000L, timeDuration.currPrecisionDuration);
 
     timeDuration = DateTimeUtils.constructTimeDuration("1y");
     Assert.assertEquals(12, timeDuration.monthDuration);
-    Assert.assertEquals(0, timeDuration.nsDuration);
+    Assert.assertEquals(0, timeDuration.currPrecisionDuration);
 
     timeDuration = DateTimeUtils.constructTimeDuration("1mo");
     Assert.assertEquals(1, timeDuration.monthDuration);
-    Assert.assertEquals(0, timeDuration.nsDuration);
+    Assert.assertEquals(0, timeDuration.currPrecisionDuration);
 
     timeDuration = DateTimeUtils.constructTimeDuration("1y1mo");
     Assert.assertEquals(13, timeDuration.monthDuration);
-    Assert.assertEquals(0, timeDuration.nsDuration);
+    Assert.assertEquals(0, timeDuration.currPrecisionDuration);
   }
 }
